Needed to reorganize things so I could loop keeping training with various learning rates and discounts.  LR=0.2 and DIS=0.4 seemed to be the "best" for the rapid ascent phase (300_000 episodes). In terms of the code, keeping_learning.py does some init work, and now calls the newly created keeping_train.py, which is actually the guts of the training (which used to be in keeping_learning.py).  In keeping_train.py you have now a new classe there: class KeepingTrain.
Also I improved solo_play.py with some error checking, and also querying of the scoring table to see what it would do when it's time to score.  BTW it's not bad, but I found an anomaly with roll 6-6-6-6-6.  It scored 0 in the fives, which obviously is a dumb move.

Also for the masking of action 60, I masked out a few actions for rolls that are not actually one dice away from the straight.