Applied the uniform probability rolling (i.e. rolling any of the 252  entries from the dice roll dictionary.  Recall that this means that '33333' has an equal probability of being rolled as '12334'.  Retrained the scoring table (I think for 25M episodes).  Added the necessary code in "scoring_learning.py".  Added function "seed_score_table.py" to initialize the scoring table with some random categories marked as unavailable.  All this in an effort to do like the keeping learning (with uniform rolling and randomly set score tables.  And, the end goal of this is to have more uniform visitation of the q_table rows.
Unfortunately I didn't check if that happened with the scoring table during this last training, but, in this check-in, I add a nice plot of the row visit rate for the keeping q_table.  It was a 25M training.  The plot has two sub plots, the first is the row visit which has a very nice and long flat zone (in log space at least).  Second plot is the streak suration.  This gives a bit of an indication about the q table convergence.  A "streak" is how long has a row been visited without its winning action changing.  You want that number to be large w.r.t the number of times the row has been visited.  That chart is not as swell as the previous.  To see this look at the right extreme, it breaks down in the last part.  Not shown in the chart is that you have zeros.  You have about 15_000.